Remove transfers slice, update IBC and Transfer send forms
