+++ apps/namada-interface/src/App/Token/IBCTransfer/index.ts
@@ -1,0 +1,1 @@
+export { default as IBCTransfer } from "./IBCTransfer";
