--- apps/namada-interface/src/App/Token/Transfers/index.ts
@@ -1,2 +1,0 @@
-export { default as Transfers } from "./Transfers";
-export { default as TransferDetails } from "./TransferDetails";
