+++ apps/namada-interface/src/App/Token/index.ts
@@ -1,1 +1,4 @@
-export { default as TokenDetails } from "./TokenDetails";
+export * from "./TokenSend";
+export * from "./TokenReceive";
+export * from "./IBCTransfer";
+export * from "./types";
