+++ apps/namada-interface/src/slices/index.ts
@@ -1,5 +1,4 @@
 export { default as accountsReducer } from "./accounts";
-export { default as transfersReducer } from "./transfers";
 export { default as channelsReducer } from "./channels";
 export { default as settingsReducer } from "./settings";
 export { default as coinsReducer } from "./coins";
